change: move selector stuff out of the api package
